+++ utils/handlers/EntryInput.ts
@@ -3,7 +3,6 @@ import { useState } from 'preact/hooks';
 import { useTagList } from 'hooks';
 import { bring, isURL } from 'utils';
 import { iEntryInput } from 'islands/EntryInput/index.tsx';
-import { updateEntryList } from 'signals';
 import { iEntry } from 'db/entry.ts';
 
 type Steps = 'entrymark' | 'tags';
@@ -23,15 +22,18 @@ export default function (props: iEntryInput) {
       ? `/api/entries/${entry._id}/update`
       : '/api/entries/new';
 
-    await bring<iEntry>(setApiUrl, 'POST', {
-      utc_created_at: entry.utc_created_at,
-      content: entryValue,
-      tags: tags,
-      entry_mark: entryMark,
-      day_count: entry.day_count,
-    }, 'Create entry error.');
-
-    updateEntryList.value++;
+    await bring<iEntry>(
+      setApiUrl,
+      'POST',
+      {
+        utc_created_at: entry.utc_created_at,
+        content: entryValue,
+        tags: tags,
+        entry_mark: entryMark,
+        day_count: entry.day_count,
+      },
+      'Create entry error.',
+    );
 
     setEntryValue('');
     setEntryMark('');